requirements.txt
aiohttp>=3.8
aioredis>=2.0
async-timeout
pytest
locust
fastapi
uvicorn
pyyaml
